test(maz-ui): fix tests
